Moved PC Class to new file. Switch case routes.
